tester: don't send empty message for empty version suffix
The desktop app currently crashes when handling a testRes message with an
empty string.

That bug should be fixed over there, but it turns out there's a bug in the
firmware as well, in that the suffix is not prefixed by a "-" as it should be.

Adding this prefix, and making sure we don't send anything if there is
no suffix, fixes that and "accidentally" works around the desktop bug.