Fix clunking sound from solenoids
The "clunking" sound that is occasionnally heard when using AYAB
is caused by all solenoid armatures being released simultaneously
when all solenoids are turned off.

This does not happen with Brother electronics because they default
to keeping solenoids on.

This commit changes AYAB's behavior to match that behavior of the
Brother electronics. Note that this required abandoning the third-party
library used to drive the I/O expanders (Adafruit_MCP23008) since
part of its non-skippable initialization code was forcing all outputs
to off. It turns out the amount of code needed to directly drive the
MCP23008 expanders using the Arduino Wire library is only a dozen
lines of code anyway.

Fixes #15.