Restart when ayab-desktop start a new session
Legacy platforms rely on hardware reset when ayab-desktop
start a new session and open the serial port but this doesn't
work on platforms like UNO R4.

To remain compatible with current ayab-desktop implementation
a reset of the knitter state-machine is issued upon reception
of a reqInit message while in Operate mode.