Encoders: only look at leftmost edge of left sensor signal
The code in encA_rising used to reset the carriage position to "zero" for
every needle where the sensor was triggered.
This caused the zero position to be taken as the rightmost position where
a magnet was detected (a magnet is usually detected for two to three
needle widths).
In turn this caused solenoids to sometimes be activated too late to correctly
select needles.
By checking for the signal going from "no magnet detected" to "any
magnet detected", the "zero" position will be set only at the leftmost edge
of the magnet detection area. This should shift everything left by one or two
needles, leaving more time for the solenoids to be activated.